// src/pumpsdk/generated/accounts/bondingCurve.ts
export type BondingCurve = {
  discriminator: ReadonlyUint8Array;
  virtualTokenReserves: bigint;
  virtualQuoteReserves: bigint;
  realTokenReserves: bigint;
  realQuoteReserves: bigint;
  tokenTotalSupply: bigint;
  complete: boolean;
  creator: Address;
  isMayhemMode: boolean;
  isCashbackCoin: boolean;
  quoteMint: Address;
  creatorFeeBps: bigint;
  canEditCreatorFee: boolean;
  isHolderReward: boolean;
};